Forgot to tick other files in the commit + push

+++ datapoints.h
@@ -4,7 +4,7 @@
 #include <bitset>
 #include <cstdio>
 
-typedef float value_type2; // or are we writing data of type double?
+typedef double value_type2; // or are we writing data of type double?
 
 // load data from a binary file into arrays x and y 
 // for layout, look at commentaries inside function body
